fix(work-capsules): satisfy WS9 policy gates — extract liveness inventory, plan coverage receipt
Green-up for PR #4056 (WS9 / BI-CBAAEA94):
- Module Size Guard: extract the list_work_capsules liveness join/annotation into
  apps/web/lib/work-capsules/liveness-inventory.ts so mcp-handlers.ts returns to
  868 LOC (<= its 884 baseline; ratchet only allows shrink). Tool behaviour
  unchanged (16 work-capsules tool tests still pass).
- Plan Backlog Coverage: recorded an atomic coverage receipt via
  record_plan_backlog_coverage and copied it into the plan (Decision/Parent/
  Receipt/Rationale/Dependencies).

Docs-Impact-Decision: internal changes only — work-capsules-pack.ts adds a staleOnly param + a liveness note to an existing tool description, and taskrun-watchdog.ts adds a governed reaper hook; neither changes a user-guide page.
Seed-Fit-Decision: global-default
